+++ src/components/CartItemContainer/CartItemContainer.tsx
@@ -2,7 +2,7 @@ import * as S from './style';
 
 import CartItem from '../CartItem/CartItem';
 import CheckBox from '../CheckBox/CheckBox';
-import { cartItemsState } from '../../recoil/selectors';
+import { cartItemsState } from '../../recoil/atoms';
 import { removeCartItem } from '../../apis';
 import useCheckedItemIds from '../../hooks/useCheckedItemIds';
 import useItemQuantity from '../../hooks/useItemQuantity';
