+++ src/hooks/useItemQuantity.ts
@@ -1,4 +1,4 @@
-import { cartItemsState } from '../recoil/selectors';
+import { cartItemsState } from '../recoil/atoms';
 import { updateCartItemQuantity } from '../apis';
 import { useRecoilState } from 'recoil';
 
@@ -13,11 +13,17 @@ export default function useItemQuantity() {
 
   const increaseQuantity = async (id: number) => {
     const nextCartItems = [...cartItems];
-    const targetItem = nextCartItems.find((item) => item.id === id);
-    if (targetItem === undefined) return;
-    await updateCartItemQuantity(id, targetItem.quantity + 1);
-    targetItem.quantity++;
+    const targetIndex = nextCartItems.findIndex((item) => item.id === id);
+    if (targetIndex === -1) return;
+    nextCartItems[targetIndex] = {
+      ...nextCartItems[targetIndex],
+      quantity: nextCartItems[targetIndex].quantity + 1,
+    };
     setCartItems(nextCartItems);
+    await updateCartItemQuantity(id, nextCartItems[targetIndex].quantity + 1).catch(() => {
+      alert('네트워크 접속이 불안정합니다. 다시 시도해주세요');
+      setCartItems(cartItems);
+    });
   };
 
   const decreaseQuantity = async (id: number) => {
@@ -25,9 +31,12 @@ export default function useItemQuantity() {
     const targetItem = nextCartItems.find((item) => item.id === id);
     if (targetItem === undefined) return;
     const nextTargetQuantity = Math.max(0, targetItem.quantity - 1);
-    await updateCartItemQuantity(id, nextTargetQuantity);
     targetItem.quantity = nextTargetQuantity;
     setCartItems(nextCartItems);
+    await updateCartItemQuantity(id, nextTargetQuantity).catch(() => {
+      alert('네트워크 접속이 불안정합니다. 다시시도해주세요');
+      setCartItems(cartItems);
+    });
   };
 
   return { getQuantity, increaseQuantity, decreaseQuantity };
